@@ -30,14 +30,16 @@
     "chart.js": "3.4.1",
     "chroma-js": "2.4.2",
     "formik": "^2.2.9",
+    "konva": "^8.3.10",
     "prop-types": "15.8.1",
-    "react": "17.0.2",
+    "react": "^18.2.0",
     "react-bootstrap-sweetalert": "^5.2.0",
     "react-chartjs-2": "3.0.4",
     "react-countup": "6.1.1",
-    "react-dom": "17.0.2",
+    "react-dom": "^18.2.0",
     "react-flatpickr": "3.10.7",
     "react-github-btn": "1.2.1",
+    "react-konva": "^18.1.1",
     "react-query": "^3.39.1",
     "react-router-dom": "6.2.1",
     "react-scripts": "5.0.0",
