@@ -6,9 +6,9 @@ namespace Vayosoft.Core.SharedKernel.Models.Pagination
     public class PagedEnumerable<T> : IPagedEnumerable<T>
     {
         private readonly IEnumerable<T> _inner;
-        private readonly int _totalCount;
+        private readonly long _totalCount;
 
-        public PagedEnumerable(IEnumerable<T> inner, int totalCount)
+        public PagedEnumerable(IEnumerable<T> inner, long totalCount)
         {
             _inner = inner;
             _totalCount = totalCount;
